fix(metrics): move phone usage module to metrics
